Configure the swiftlint for support Xcode 15.
Consideration for Xcode 15.0
Xcode 15 made a significant change by setting the default value of the ENABLE_USER_SCRIPT_SANDBOXING build setting from NO to YES. As a result, SwiftLint encounters an error related to missing file permissions, which typically manifests as error: Sandbox: swiftlint(19427) deny(1) file-read-data.

To resolve this issue, it is necessary to manually set the ENABLE_USER_SCRIPT_SANDBOXING setting to NO for the specific target that SwiftLint is being configured for.

// Ref: https://github.com/realm/SwiftLint